+++ Packages/com.unity.inputsystem/InputSystem/Editor/Analytics/InputBuildAnalytic.cs
@@ -195,6 +195,9 @@ namespace UnityEngine.InputSystem.Editor
                 feature_paranoid_read_value_caching_checks_enabled =
                     settings.IsFeatureEnabled(InputFeatureNames.kParanoidReadValueCachingChecks);
 
+                feature_use_imgui_editor_for_assets =
+                    settings.IsFeatureEnabled(InputFeatureNames.kUseIMGUIEditorForAssets);
+
                 feature_disable_unity_remote_support =
                     settings.IsFeatureEnabled(InputFeatureNames.kDisableUnityRemoteSupport);
                 feature_run_player_updates_in_editmode =
@@ -328,6 +331,12 @@ namespace UnityEngine.InputSystem.Editor
             public bool feature_paranoid_read_value_caching_checks_enabled;
 
             /// <summary>
+            /// Represents internal feature flag <see cref="InputFeatureNames.kUseIMGUIEditorForAssets" />
+            /// as defined in InputSystem 1.8.x.
+            /// </summary>
+            public bool feature_use_imgui_editor_for_assets;
+
+            /// <summary>
             /// Represents internal feature flag <see cref="InputFeatureNames.kDisableUnityRemoteSupport" />
             /// as defined in InputSystem 1.8.x.
             /// </summary>
@@ -376,7 +385,7 @@ namespace UnityEngine.InputSystem.Editor
 
             public void OnPostprocessBuild(BuildReport report)
             {
-                InputSystem.s_Manager?.m_Runtime?.SendAnalytic(new InputBuildAnalytic(report));
+                InputSystem.manager?.runtime?.SendAnalytic(new InputBuildAnalytic(report));
             }
         }
     }